Added Miners
Claymore - Cryptonight
EWBF - Equihash
Cgminer - Reorder
Sgminer - Genesis Mining

+++ Miners/Ccminer.ps1
@@ -5,10 +5,7 @@ if((Test-Path $Path) -eq $false)
     $FileName = "ccminer.zip"
     try
     {
-        if(Test-Path $FileName)
-        {
-            Remove-Item $FileName
-        }
+        if(Test-Path $FileName){Remove-Item $FileName}
         Invoke-WebRequest "https://github.com/nicehash/ccminer-tpruvot/releases/download/1.8-decred-nicehash-2/ccminer.zip" -OutFile $FileName -UseBasicParsing
         Expand-Archive $FileName (Split-Path $Path)
     }
@@ -21,6 +18,7 @@ if((Test-Path $Path) -eq $false)
 $Name = (Get-Item $script:MyInvocation.MyCommand.Path).BaseName
 
 $Algorithms = [PSCustomObject]@{
+    Cryptonight = 'cryptonight'
     Sia = 'sia'
     Yescrypt = 'yescrypt'
     BlakeVanilla = 'vanilla'
@@ -42,5 +40,6 @@ $Algorithms | Get-Member -MemberType NoteProperty | Select -ExpandProperty Name 
         Arguments = -Join ('-a ', $Algorithms.$_, ' -o stratum+tcp://$($Pools.', $_, '.Host):$($Pools.', $_, '.Port) -u $($Pools.', $_, '.User) -p x')
         HashRates = [PSCustomObject]@{$_ = -Join ('$($Stats.', $Name, '_', $_, '_HashRate.Day)')}
         API = 'Ccminer'
+        Port = 4068
     }
 }
